pyproject.toml
[project]
name = "spark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning",
    "pandas",
    "pyarrow",
    "torch",
    "wandb",
    "loguru",
    "einops",
    "rich",
    "tensorboard",
    "hydra-core>=1.3.2",
    "dotenv>=0.9.9",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["spark*"]

[dependency-groups]
dev = [
    "pytest",
    "huggingface-hub",
    "ipykernel",
    "jupyter",
    "jupyterlab",
    "matplotlib",
    "seaborn",
]
